Move form title inside page content to handle key events
